adjust for iPad view
