+++ firefox-ios/Client/Frontend/Library/LibraryPanelContextMenu.swift
@@ -9,7 +9,7 @@ import Shared
 protocol LibraryPanelContextMenu {
     func getSiteDetails(for indexPath: IndexPath) -> Site?
     func getContextMenuActions(for site: Site, with indexPath: IndexPath) -> [PhotonRowActions]?
-    func getShareAction(site: Site, sourceView: UIView, coordinator: LibraryPanelCoordinatorDelegate?) -> PhotonRowActions
+    func getShareAction(site: Site, sourceView: UIView, delegate: LibraryPanelCoordinatorDelegate?) -> PhotonRowActions
     func presentContextMenu(for indexPath: IndexPath)
     func presentContextMenu(
         for site: Site,
@@ -106,12 +106,12 @@ extension LibraryPanelContextMenu {
         return [openInNewTabAction, openInNewPrivateTabAction]
     }
 
-    func getShareAction(site: Site, sourceView: UIView, coordinator: LibraryPanelCoordinatorDelegate?) -> PhotonRowActions {
+    func getShareAction(site: Site, sourceView: UIView, delegate: LibraryPanelCoordinatorDelegate?) -> PhotonRowActions {
         return SingleActionViewModel(
             title: .ShareContextMenuTitle,
             iconString: StandardImageIdentifiers.Large.shareApple) { _ in
                 guard let siteURL = URL(string: site.url, invalidCharacters: false) else { return }
-                coordinator?.shareLibraryItem(url: siteURL, sourceView: sourceView)
+                delegate?.shareLibraryItem(url: siteURL, sourceView: sourceView)
         }.items
     }
 }
